requirements.txt
# ML Libraries
tensorflow-gpu>=2.12.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
scikit-learn>=1.2.2
xgboost>=1.7.5
lightgbm>=4.0.0
optuna>=3.1.0
numpy>=1.23.5
pandas>=2.0.0
pandas_ta>=0.3.14b0
scipy>=1.10.0

# CUDA Dependencies
nvidia-cuda-toolkit>=12.0
nvidia-cudnn>=8.9

# Trading and Analysis
python-binance>=1.0.17
python-dotenv>=1.0.0
ccxt>=1.60.0

# Dashboard
dash>=2.14.1
dash-bootstrap-components>=1.5.0
plotly>=5.18.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.18.0

# Utilities
asyncio>=3.4.3
aiohttp>=3.8.4
python-telegram-bot>=20.0
joblib>=1.2.0
psutil>=5.9.0
PyYAML>=6.0

# Testing
pytest>=7.4.0 